Add type signature fallback and JIT objfile restriction.
To support type signature fallback, the lookup_signatured_type method is extended.
If no type signature was found in the current objfile or in a dwo file, the dwraf-type-signature-fallback option is checked:
'off': return nullptr
'main': only check the current main objfile for the type signature.
This reduces the additional load for a fallback as only one additional objfile is checked.
The main objfile often contains major parts of the debug info, thus it is also likely that the requested type signature is found there.
If the main objfile is linked to dwo/dwp files, those are also checked for the type signature.
This should be the same behavior as if the type signature is lookup up without fallback.
'full': check all objfiles in the current prospace.
Requires potentially more time than just checking the main objfile, but it will also find a type signature in a loaded shared library.

If a type signature was found in a fallback objfile, the per_objfile after lookup_signatured_type is the fallback per_objfile.
This ensures that if the type unit still needs to be read, it is read from the correct objfile.

Additionally, a flag is added to objfiles that marks an objfile as a JIT objfile.
With this information the type signature fallback can be restricted to only work for JIT objfiles.
This is a separate option because JIT objfile are self-contained but usually share debug info (especially type information) with some other already loaded objfile.
So this leaves loading other objfiles as is to avoid unexpected behavior if needed.