Add new commands for controlling type signature fallback.
'set dwarf-type-signature-fallback (off|main|full)':
Enable type signature fallback.
If a type signature is not found in the current objfile, either looks up the 'main' objfile or all objfiles in a progspace to find the type signature.
Usually a type signature is unique, ther should only be one type unit for each type signature.
Therefore, it should be okay to check other objfile for a type signature.
However, whether this is really the case depends on the debug info producer.
This is why its disabled by default.

'set dwarf-type-signature-fallback-jit':
Whether the type signature fallback is enabled for only JIT objfiles.
JIT objfiles are self-contained, but usually share most debug info with some already loaded objfile.
This is why JIT objfiles are a major use case for the type signature fallback.
This option is added to restrict the type signature fallback to match this exact use-case and provides a more granular use of the type signature fallback.